Use Repository.defaultBranchRef to find latest commits for each repo
This seems to cause more repos to be counted (13 more) for some reason. I
think the previous query for repos was effectively filtering out these
repos. I think this is because the previous way of fetching the most recent
ref for repos _wasn't_ fetching the most recent ref. For example, it thought
that the latest commit for houseabsolute/perl-code-tidyall was in 2018, even
though I made a commit to the master branch in 2022.

Using defaultBranchRef seems to be finding the right refs.